Escape management in only one method
I've moved the leave full screen on escape from the media_viewer to the
existing escape method in state.rs to have all the behaviour in only one
place.

Doing this I've also added the functionality to go back to the chat when
we're in the MediaViewer and to do this I've to modify a little the
autocomplete, to avoid the close when we press escape on the
autocomplete popover.

I've also changed the key management on key press instead of key release
because the keypress is grabbed by child windows or popovers and the
release isn't.

Close #277