Improve avatar async loading
The avatar was not redrawing after image download, I've moved the avatar
loading retry from the drawing method to a separated one to call just
after the backend respond so we don't retry if it's not needed and we
can redraw when the image is ready.

I've also changed the letter-avatar git dependency to use the crates.io
one.

We should improve this async user data loading, this will change in the
near future when we change to the new data model and we can have a good
cache system. Meantime this solution works.