Fix Improper Decoration Layout For HeaderBar
After commit 1db7847882113e8041e170a260c6150dd2169694
Fractal would display window decorations incorrectly
if decorations were set to be on the left, showing on the
inner edge of the split. This is because window decorations
were only set for the right HeaderBar. This commit
makes Fractal detect the GTK setting and set the layout
correctly.