Remove the custom strn! macro.
This custom macro adds extra complexity since the user can't
know what it does and has to be familiar with it. It's behavior is
trivial and it does not justify a custom macro for a one-liner.

This also has the benefit of reducting the compile time, since its
now replace by a funtion call and the compiler no longer needs
to expand each occurence of the macro.