message.rs: Use itertools to group quotes lines
I've tried to simplify the code that group lines by quote/no-quote, it's
a lot simpler using the itertools crate.

This patch also adds a new enum for the message part, instead of use a
simple bool, so we can manage other kind of rendering in a message. I
was thinking about the Markdown code block for example.