Initial sync speed up
The stickers loading was done syncronously so the initial sync wasn't
launched after the stickers was loaded and that increments a lot the
initial sync time.

I've also removed the m.receipt from the filter and use the m.fully_read
that comes with the initial sync to know the last read message.