MessageMenu: Copy selected text action to menu
The message menu override the default label context menu and we can't
copy selected text using the label context menu so I've added a copy
selected text option to the message menu to do the same.

We can do something similar to add two new actions to this menu:

 * Copy link
 * Open link

We can try to check if we've a link below the cursor and in that case we
can show these two new actions.

See #195