notify: avoid clone all rooms for the notification
This patch moves the base summary generation outside of the closure to
avoid the clone of all rooms. The rooms struct can have a lot of info
and we can avoid to clone int this case if we build the string outside
the closure and then we replaces the '{user}' string with the real
username when we've it.