cache: Use mdl cache instead of json
With this change we're using lmdb as the cache storage and it works.

In this case we're not using the AppMsg cache model to store messages,
we're storing all the room with all the msgs.

In the future will be better to split that in two, so it'll be faster to
add, delete or update a room message, because with this version any of
those operations requires to serialize the whole message array again.