fractal-gtk: Use FileChooserNative and set default location
In order for saving files to work without poking holes in the
flatpak sandbox, FileChooserNative needs to be used for portal
support. As a separate UX issue, the filechooser defaulted to
home for downloads, when ~/Downloads would have been more
appropriate. This was fixed alongside the FileChooserNative
changes.

Related to https://gitlab.gnome.org/World/fractal/issues/343