Send info data for image attachment
This patch adds the ability to extend a message event content with
custom json and use that new functionality to attach the image info for
image messages.

I've done this way becase in the future the protocol can support other
fields and we can add that from the client side without the need of the
backend to difference between that.

Fix #348