fractal-gtk: Use HdyLeaflet for mobile UI
HdyLeaflet is a widget that can be used in split-header apps
like Fractal to split the panes into views, one primary and one
secondary.

This commit splits the main view of Fractal with HdyLeaflet so
at small sizes the room list is the primary view, and the room
history is the secondary view.