uibuilder.rs: Add scroll_widget to uibuilder
A component previously used in main_window.ui  was moved into
scroll_widget.ui and called from App.ui.builder. However with
the move scroll_widget wasn't added to the rest of the files
in the GtkBuilder used for that call, causing a panic on
startup.

Now uibuilder adds the resource for scroll_widget.ui