notify: Use GBytesIcon
Previously we sent a GFileIcon for notifications, but the
flatpak portal does not support GFileIcon (https://github.com/flatpak/xdg-desktop-portal/issues/317).
Because of this notifications from the flatpak were
being dropped.

Now we create a GBytesIcon by loading a GFile and reading
the bytes, and set that as the icon for notifications.

Closes https://gitlab.gnome.org/GNOME/fractal/issues/490