Fix typing notification margin
The margin of the typing_label was not exactly aligned with the text, we
have the padding left of the list (18/6) plus the size of the avatar
(40) plus double of the padding of the row (9 * 2).

With the handy column the list left padding changes from 18 to 6 so we
need to control that case and change the margin according to that.

Fix https://gitlab.gnome.org/GNOME/fractal/issues/492