api: Log every channel sender send error
We were ignoring channel send errors and it could be interesting to have
a log about this not sent messages.

This patch adds a new trait to extend the Result<(), SendError<T>> type
and provide a new method called expect_log, that doesn't panic on error
but logs the message in that case.