Make entry editable by default when no power_levels
For new rooms we're not receiving power_levels event, so we need to
assume that the user can write.

This is a quick fix to allow edit new rooms, but it could be better to
request power_levels to the backend when we've no this information so we
can disable the text entry if the user has no permissions to do that,
but without the information we should assume that the user is able to
write because it's the usual case.

Fix https://gitlab.gnome.org/GNOME/fractal/issues/539