login: Don't remove the Homeserver schema unconditionally.
The current Homeserver URL parsing logic unconditionally removes the URL
schema if it's either "http://" or "https://", then prepends
the "https://" schema back immediately after.

While this correctly adds a "https://" schema for an URL without a
schema, it disallows connections to a non-TLS host. Synapse, rightfully
so, disallows non-TLS connections. Connecting to such a host might be
desirable anyway to connect to a logging/debugging reverse proxy such as
mitmproxy or to connect to Pantalaimon, a reverse proxy that enables
client-transparent E2E encryption.

Closes: https://gitlab.gnome.org/GNOME/fractal/issues/520