Image: Don't try to load thumbs from non remote paths
Images sent from fractal was using the local path as thumbnail,
something like /tmp/fractal_RANDOM.png, that's the reason why the image
thumbs are not shown in the history.

This patch don't try to load thumbnails that are not mxc://,
http or https and will fallback to the msg full image instead.

Fix https://gitlab.gnome.org/GNOME/fractal/issues/572