Fix how to get the duration of audio/video
Before, the gtk main thread sometimes froze when sending an audio or video
attachment. That was due to trying to get the duration of the audio/video
using gstreamer_editing_services. Now, using gstreamer_pbutils,
that doesn't happen anymore.

Furthermore, when creating an audio player in the room history, the
duration of the audio was displayed as 00:00 until play was hit. Now,
the duration of the audio gets displayed correctly from the moment of the
creation of the player on.