Fix case handling when not finding file uri
For files (pdf, image, audio, video) in the room history and media viewer,
the backend is asked to provide the uri of that file. Before, when the
backend failed to find out the uri, an empty string was sent back to the
frontend. That triggered unexpected behaviour such as initializing an
empty audio/video player or not opening the file chooser when hitting
download or save as.

With this commit, errors finding a file uri get handled separately.