issue template: Update bug report template to include useful info
Our current bug template is almost the same as our feature request template. This leads to users not always providing the information we need to act on a bug.

This commit updates the template so that users will
be able to provide more detailed bug reports without blank
or placeholder lines.