Support matrix servers in a subdirectory.
Previously, the user-provided URL was join'ed with an absolute path
(e.g. "/_matrix/..."). Thus, the path was replaced with the given one.
In other words, fractal assumed that matrix servers are always available
at "/".

This commit appends a relative path, so that matrix servers in
subdirectories are supported as well.