directory: Use server_chooser_popover just once
It's not recommended to use the same gtk::Popover object attached to two
different gtk::MenuButton because the popover has a parent and has an
attached to to place it correctly and having this object attached to two
different buttons causes some misplacements.

This patch changes the layout to use the HdySqueezer inside the
MenuButton, this way we only have one menu button with two layouts
instead of two different MenuButtons.

This should fix the wrong placement of the directory popover.