media viewer: Fix image sizing
Previously, images in the Media Viewer would not be resized dynamically if the
window was resized. They would stay at the size they got when they were loaded.

Fixed by changing the behavior of the image widget. `fit_to_width` is renamed
`shrink_to_fit` to better reflect this new behavior. When it is enabled, the
image widget does not change its size request. Instead, it always resizes the
image to fit within its allocation, shrinking (but never expanding) it if
necessary.

Fixes #493.