fixed strict comparison of f64
this is an interesting error as it helps to fix floating point errors
Comparing 2 f64 values should always be done with greater an value or lower an value as
the implementation of floating point values are probably never exact a specific value.
Note: This are errors for clippy - not warnings

This is what clippy says for that:
error: strict comparison of `f32` or `f64`
   --> fractal-gtk/src/widgets/scroll_widget.rs:315:38
    |
315 |                 if now < end_time && adj.get_value().round() != end.round() {
    |                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider comparing them within some error: `(adj.get_value().round() - end.round()).abs() > error`
    |
note: `std::f32::EPSILON` and `std::f64::EPSILON` are available.
   --> fractal-gtk/src/widgets/scroll_widget.rs:315:38
    |
315 |                 if now < end_time && adj.get_value().round() != end.round() {
    |                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#float_cmp