gtk: Manage all app arguments with clap
This patch enables the clap arguments for all compilations, debug or
production.

We're managing options with clap so we shouldn't pass the arguments to
the appliation to avoid problems with arguments managed by clap, for
example for calls with -vv, that's an unknown option for the
gio::Application.

Fix https://gitlab.gnome.org/GNOME/fractal/-/issues/639