fractal-gtk: Add AppOp to AccountSettings
During the App::connect_gtk the login_data is None, so the
AccountSettings is not initialized correctly.

We need to get the login_data when we are going to change the avatar,
because that data can change during the app usage.

This patch uses the AppOp shared reference to get the updated data just
on click.

Fix https://gitlab.gnome.org/GNOME/fractal/-/issues/652