directory: Fix search for other homeservers
The matrix server request expect the "server" parameter to be a string.
If the "server" in Parameters is "Option<Host<String>>" it serializes to
somethign like 'Domain("matrix.org")'.

This patch makes the server parameter a string so the request is built
correctly everytime. The "Host" enum is converted to an string during
the parsing.