Fix join to room by alias
The join_room method parameter was a RoomId so it was not possible to
join to a room by alias. This patch changes that parameter to a
RoomIdOrAliasId so we can use a RoomId or an AliasId.

This patch also encode the `#` in the AliasId because in other case, the
url is not correct and the request will fail.

Besides the above fixes, this patch also improves a bit the error
management for the join_room method, so the error dialog will have more
information, for example if the room is not found on the server or
something like that. This information is the direct response from the
matrix server, but I think it's okay to show that to the user as
feedback because in other case the user should run Fractal in debug mode
to know what's happening.

Fixes https://gitlab.gnome.org/GNOME/fractal/-/issues/658