room_settings: Round corners, adjust spacing
A newer pattern for GNOME is rounded list corners.
In GTK4 GtkFrame automatically has rounded corners.
For now, we can adopt this style pattern by rounding
our frame and it's contents manually.

This commit also adjusts the spacing between the search
and room member list, for a less cramped feeling.