Refactor message menu
Previously our message menu had a strange API where we used
the pointer of the screen instead of the coordinates of
the relevant event, and automatically showed on "new".

This commit refactors the message menu to use the coords
of the event and only show using an explicit "show_at_coords()" function if not triggered
by a GtkMenuButton.

Fixes https://gitlab.gnome.org/GNOME/fractal/-/issues/555