Message Widget: Make ol/ul a list of boxes
Instead of adding a bullet to each label, create a horizontal box with a
bullet and the text, this way the text is properly spaced after
wrapping, i.e.

* some wrapped
text

* some wrapped
  text

where the second is the implemented and correct behavior. Additionally,
the bullet cannot be selected as part of the message.